STYLE: Re-organized file structure
